backend/requirements.txt
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
astroid==2.11.6
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.1.0
bleach==4.1.0
blinker==1.5
boto3==1.26.44
botocore==1.29.44
braintree==4.20.0
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.5
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.6
Django==4.1.4
django-cors-headers==4.0.0
django-filter==23.2
django-multiselectfield==0.1.12
djangorestframework==3.14.0
dnspython==2.2.1
email-validator==1.3.0
entrypoints==0.3
executing==0.8.2
filelock==3.9.0
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.3
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.1
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
lazy-object-proxy==1.7.1
lxml==4.7.1
manage.py==0.2.10
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.7.0
mistune==0.8.4
mypy-extensions==0.4.3
mysqlclient==2.1.1
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pipenv==2023.2.18
platformdirs==2.4.1
prometheus-client==0.13.1
prompt-toolkit==3.0.26
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pylint==2.14.3
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
pytz==2023.3
pyzmq==22.3.0
requests==2.28.0
s3transfer==0.6.0
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
sqlparse==0.4.3
stack-data==0.1.4
terminado==0.13.1
testpath==0.5.0
tomli==2.0.0
tomlkit==0.11.0
tornado==6.1
traitlets==5.1.1
urllib3==1.26.9
virtualenv==20.21.0
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
